Extract plug-in install code to own class
